move ManifestFileMixin to resolver/params.hh
